+++ app/Http/Controllers/Settings/SettingsController.php
@@ -27,6 +27,9 @@ class SettingsController extends Controller
     private static $fields = [
         'common.logo_file' => \App\Settings\Common\LogoFile::class,
         'common.signet_file' => \App\Settings\Common\SignetFile::class,
+        'common.favicon_32_file' => \App\Settings\Common\Favicon32File::class,
+        'common.favicon_180_file' => \App\Settings\Common\Favicon180File::class,
+        'common.favicon_192_file' => \App\Settings\Common\Favicon192File::class,
         'accounting.transactions.categories' => \App\Settings\Accounting\TransactionCategories::class,
         'accounting.transactions.projects' => \App\Settings\Accounting\TransactionProjects::class,
         'bank.undo_coupon_handout_grace_period' => \App\Settings\Bank\UndoCouponHandoutGracePeriod::class,