+++ app/Settings/Badges/LogoFile.php
@@ -8,6 +8,9 @@ use Illuminate\Support\Facades\Gate;
 
 class LogoFile extends BaseSettingsField
 {
+    private const MAX_WIDTH = 800;
+    private const MAX_HEIGHT = 800;
+
     public function section(): string
     {
         return 'badges';
@@ -51,7 +54,7 @@ class LogoFile extends BaseSettingsField
     public function setter($value)
     {
         $image = new ImageResize($value->getRealPath());
-        $image->resizeToBestFit(800, 800, true);
+        $image->resizeToBestFit(self::MAX_WIDTH, self::MAX_HEIGHT);
         $image->save($value->getRealPath());
         return $value;
     }